Wind direction solved
