+++ frontend/src/Components/HeroSection/Hero.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Video from '../../videos/video1.mp4';
 import { HeroContainer, HeroBg, MenuWrapper, MyVideo } from './HeroElements';
 
-import SwiperMenu from '../SwiperMenu';
+import SwiperMenu from '../SwiperMenu/MenuCards';
 
 const HeroSection = () => {
   return (
